Refactoring ‘InvestigativeRole’ in Participant to general ‘Role’
Combines all the classes in the SCORO ontology describing Authorship, Data, Financial, Investigative, Organisational and Project roles.

The value of each enumerator item is now a dict with the persistent URLs to each ontology member and it’s parent class, plus a title and description for human use.

The Participant schema has been updated to understand an enumerated field and return the relevant item’s value (rather than just it’s name).